wsgi is working

+++ lmdo/utils.py
@@ -3,6 +3,8 @@ import errno
 import sys
 import fnmatch
 import zipfile
+import re
+import site
 
 from lmdo.oprint import Oprint
 
@@ -21,7 +23,7 @@ def mkdir(path, mode=0777):
         return True
     return True
 
-def zipper(from_path, target_file_name, exclude=None):
+def zipper(from_path, target_file_name, exclude=None, delete_exist=True, replace_base_path=None):
     """
     Create zipped package
 
@@ -38,38 +40,51 @@ def zipper(from_path, target_file_name, exclude=None):
         compression = zipfile.ZIP_STORED
 
     #delete existing file before writing a new one
-    try:
-        os.remove(target_file_name)
-    except OSError:
-        pass
-
-    zip_file = zipfile.ZipFile(target_file_name, 'w', zipfile.ZIP_DEFLATED)
+    if delete_exist:
+        try:
+            os.remove(target_file_name)
+        except OSError:
+            pass
+    
+    mode = 'a' if not delete_exist else 'w'
+    zip_file = zipfile.ZipFile(target_file_name, mode, zipfile.ZIP_DEFLATED)
    
     Oprint.info('Start packaging directory {}'.format(from_path), 'lmdo')
    
     for root, dirs, files in os.walk(from_path):
-        if not exclude:
-            for file in files:
-                zip_file.write(os.path.join(root, file))
-        else:
-            for file in files:
-                excl = False
+        bp = root
+        if replace_base_path:
+            for p_th in replace_base_path:
+                if fnmatch.fnmatch(root, '*'+p_th.get('from_path')+'*'):
+                    bp = root.replace(p_th.get('from_path'), p_th.get('to_path'))
 
+        if not exclude:
+            for f in files:
+                zip_file.write(os.path.join(root, f), os.path.join(bp, f))
+        else:
+            for f in files:
+                excl = False
                 #check if file/folder should be excluded
-                if exclude['dir']:
+                if exclude.get('dir'):
                     for ex_dir in exclude['dir']:
                         if fnmatch.fnmatch(root, ex_dir):
                             excl = True
                             break
 
-                if exclude['file']:
+                if exclude.get('file'):
                     for ex_file in exclude['file']:
-                        if fnmatch.fnmatch(file, ex_file):
+                        if fnmatch.fnmatch(f, ex_file):
                             excl = True
                             break
 
-                if not excl:
-                    zip_file.write(os.path.join(root, file))
+                if exclude.get('file_with_path'):
+                    for ex_file in exclude['file_with_path']:
+                        if fnmatch.fnmatch(os.path.join(root, f), ex_file):
+                            excl = True
+                            break
+
+                if not excl:                    
+                    zip_file.write(os.path.join(root, f), os.path.join(bp, f))
 
     zip_file.close()
 
@@ -109,4 +124,28 @@ def sys_pause(message, match):
     if not fnmatch.fnmatch(name, match):
         Oprint.err('Exit excecution', 'lmdo')
 
- 
+def update_template(content, to_replace):
+    """Replace template content"""
+    for k, v in to_replace.iteritems():
+        content = content.replace(k, v)
+    
+    return content
+
+def get_template(name):
+    """Find template from package"""
+    found_dir = False
+    pkg_dir = site.getsitepackages()
+    for pd in pkg_dir:
+        if os.path.isdir(pd + '/lmdo'):
+            found_dir = '{}/lmdo/local_template/{}'.format(pd, name)
+            if os.path.isfile(found_dir):
+                break
+            else:
+                found_dir = False
+    
+    if not found_dir:
+        Oprint.warn('Template file {} is missing'.format(name), 'lmdo')
+
+    return found_dir
+
+
